+++ service/src/main/java/org/databiosphere/workspacedataservice/dao/DataSourceConfig.java
@@ -6,9 +6,6 @@ import org.springframework.boot.jdbc.DataSourceBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
-import org.springframework.integration.jdbc.lock.DefaultLockRepository;
-import org.springframework.integration.jdbc.lock.JdbcLockRegistry;
-import org.springframework.integration.jdbc.lock.LockRepository;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 
 @Configuration
@@ -25,18 +22,5 @@ public class DataSourceConfig {
   @Primary
   public NamedParameterJdbcTemplate namedParameterJdbcTemplate(DataSource dataSource) {
     return new NamedParameterJdbcTemplate(dataSource);
-  }
-
-  @Bean
-  public DefaultLockRepository defaultLockRepository(DataSource dataSource) {
-    DefaultLockRepository defaultLockRepository = new DefaultLockRepository(dataSource);
-    defaultLockRepository.setPrefix("sys_wds.INT_");
-    defaultLockRepository.setTimeToLive(10 * 60 * 1000); // 10 minutes
-    return defaultLockRepository;
-  }
-
-  @Bean
-  public JdbcLockRegistry jdbcLockRegistry(LockRepository lockRepository) {
-    return new JdbcLockRegistry(lockRepository);
   }
 }
